Hardware Timer Simulations & anything dependent on them are not working as intended(No Simulation Errors)Exp 19, 20, 21, 22, 23 are not working
The hardware timer simulations & anything dependent are not working and I don't know how to fix them, I tried almost everything but I don't know how to make them work, I am still trying but don't know how much will I be successful.
I tried with various CLK sources, freqencies but no luck.
This is the same reason the Hardware UART was also not working.
There are no simulation errors but the results are not satisfactory.
Exp 19, 20, 21, 22, 23 are not working.